+++ src/extract.rs
@@ -1,5 +1,6 @@
 //! Experimental extractors.
 
+pub use crate::body_hash::BodyHash;
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
 pub use crate::lazy_data::LazyData;
 pub use crate::query::Query;